scanners/osint/requirements.txt
dnspython==2.6.1
psycopg2-binary==2.9.9
pika==1.3.2
redis==5.0.4
sentinel-core @ file:///app/scanner-core
